Remove Chief-scientist from basin pages
Before you go and revert this one, know that there was a nasty almost
Heisenbug that would occur. When loading any other page before a basin
page, an error would occur when then going to a basin page. If a basin
page was the first to load (after a server restart), it would succeed.

The bug itself was traced to somewhere inside the ActiveRecord
framework and related to the additional join caused by including
cruise contacts. Since the chief scientist display on current production
cchdo has been blank for a month or two and no one has noticed, it was
decided to remove the chief scientist display. This resolves the
exception, and declutters the basin page a bit.